backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.25.2
lxml==4.9.3
dateparser==1.2.0
redis==5.0.1
celery==5.3.4 